requirements.txt
elasticsearch
apscheduler
pymongo
pykube
google-cloud-storage
pyOpenSSL
